Move tests to new unit test
